+++ src/core/package.scala
@@ -10,15 +10,10 @@ extension [A](a: A) inline def single: NonEmpty[A] = a -> List.empty[A]
 type Interval[A] = NonEmpty[A]
 type History[A]  = NonEmpty[Interval[A]]
 
-trait Page[A] extends (Mutation => Option[Mutation])
-object Page:
-  given any: Page[EmptyTuple] = Some(_)
-
-  given tuple[H, T <: Tuple](using h: Page[H], t: Page[T]): Page[H *: T] =
-    a => h(a).flatMap(t)
-
-  def apply[A](using pa: Page[A]) = pa
-end Page
+sealed trait Distance
+sealed trait Duration
+sealed trait Timestamp
+sealed trait Box
 
 type Injection[A] = (HTMLElement, A)
 trait Inject[A] extends (Injection[A] => Unit):
